[IMP] Allow to use the storage either from the payment mode or from the payment method
