--- account_payment_method_fs_storage/i18n/it.po
@@ -1,84 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_payment_method_fs_storage
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2024-05-03 17:35+0000\n"
-"Last-Translator: mymage <stefano.consolaro@mymage.it>\n"
-"Language-Team: none\n"
-"Language: it\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-"X-Generator: Weblate 4.17\n"
-
-#. module: account_payment_method_fs_storage
-#: model:ir.model,name:account_payment_method_fs_storage.model_fs_storage
-msgid "FS Storage"
-msgstr "Deposito FS"
-
-#. module: account_payment_method_fs_storage
-#. odoo-python
-#: code:addons/account_payment_method_fs_storage/models/account_payment_order.py:0
-#, python-format
-msgid "Generate and export"
-msgstr "Genera ed esporta"
-
-#. module: account_payment_method_fs_storage
-#: model:ir.model,name:account_payment_method_fs_storage.model_account_payment_method
-msgid "Payment Methods"
-msgstr "Metodi di pagamento"
-
-#. module: account_payment_method_fs_storage
-#: model:ir.model,name:account_payment_method_fs_storage.model_account_payment_order
-msgid "Payment Order"
-msgstr "Ordine di pagamento"
-
-#. module: account_payment_method_fs_storage
-#: model:ir.model.fields,field_description:account_payment_method_fs_storage.field_account_payment_method__storage
-msgid "Storage"
-msgstr "Deposito"
-
-#. module: account_payment_method_fs_storage
-#. odoo-python
-#: code:addons/account_payment_method_fs_storage/models/fs_storage.py:0
-#, python-format
-msgid "Storage is already used on at least one payment method"
-msgstr "Deposito utilizzato almeno in un metodo di pagamento"
-
-#. module: account_payment_method_fs_storage
-#: model:ir.model.fields,help:account_payment_method_fs_storage.field_account_payment_method__storage
-msgid "Storage where to put the file after generation"
-msgstr "Deposito dove salvare il file dopo la generazione"
-
-#. module: account_payment_method_fs_storage
-#: model:ir.model.fields,help:account_payment_method_fs_storage.field_fs_storage__use_on_payment_method
-msgid "Tells if storage can be set on payment methods"
-msgstr "Indica se il deposito può essere impostato in un metodo di pagamento"
-
-#. module: account_payment_method_fs_storage
-#. odoo-python
-#: code:addons/account_payment_method_fs_storage/models/account_payment_order.py:0
-#, python-format
-msgid "The file has been generated and dropped on the storage."
-msgstr "Il file è stato generato e salvato nel deposito."
-
-#. module: account_payment_method_fs_storage
-#. odoo-python
-#: code:addons/account_payment_method_fs_storage/models/account_payment_order.py:0
-#, python-format
-msgid ""
-"Unknown issue to upload the file on the storage:\n"
-"{details}"
-msgstr ""
-"Problema sconosciuto nel caricamento del file nel deposito:\n"
-"{details}"
-
-#. module: account_payment_method_fs_storage
-#: model:ir.model.fields,field_description:account_payment_method_fs_storage.field_fs_storage__use_on_payment_method
-msgid "Use On Payment Method"
-msgstr "Utilizza nel metodo di pagamento"
