--- account_payment_method_fs_storage/models/__init__.py
@@ -1,3 +1,0 @@
-from . import account_payment_order
-from . import account_payment_method
-from . import fs_storage
