--- account_payment_method_fs_storage/readme/CONTRIBUTORS.rst
@@ -1,2 +1,0 @@
-* François Honoré <francois.honore@acsone.eu>
-* Zina Rasoamanana <zina.rasoamanana@acsone.eu>
