--- account_payment_method_fs_storage/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_account_payment_method_fs_storage
