+++ account_payment_method_or_mode_fs_storage/i18n/fr.po
@@ -1,0 +1,126 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* account_payment_method_or_mode_fs_storage
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0+e\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2024-06-19 09:37+0000\n"
+"PO-Revision-Date: 2024-06-19 09:37+0000\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: account_payment_method_or_mode_fs_storage
+#. odoo-python
+#: code:addons/account_payment_method_or_mode_fs_storage/models/account_payment_mode.py:0
+#, python-format
+msgid " - disabled in method payment setting"
+msgstr " - inactif dans le cas d'une configuration sur la méthode de paiement"
+
+#. module: account_payment_method_or_mode_fs_storage
+#. odoo-python
+#: code:addons/account_payment_method_or_mode_fs_storage/models/account_payment_method.py:0
+#, python-format
+msgid " - disabled in mode payment setting"
+msgstr " - inactif dans le cas d'une configuration sur le mode de paiement"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model,name:account_payment_method_or_mode_fs_storage.model_res_company
+msgid "Companies"
+msgstr "Sociétés"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model,name:account_payment_method_or_mode_fs_storage.model_res_config_settings
+msgid "Config Settings"
+msgstr "Paramètres"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model.fields,field_description:account_payment_method_or_mode_fs_storage.field_res_config_settings__fs_storage_ids
+msgid "Fs Storage allowed"
+msgstr "Fs storages autorisés"
+
+#. module: account_payment_method_or_mode_fs_storage
+#. odoo-python
+#: code:addons/account_payment_method_or_mode_fs_storage/models/account_payment_order.py:0
+#, python-format
+msgid "Generate and export"
+msgstr ""
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model.fields.selection,name:account_payment_method_or_mode_fs_storage.selection__res_company__fs_storage_source_payment__method
+msgid "Method"
+msgstr "Méthode"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model.fields.selection,name:account_payment_method_or_mode_fs_storage.selection__res_company__fs_storage_source_payment__mode
+msgid "Mode"
+msgstr "Mode"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model_terms:ir.ui.view,arch_db:account_payment_method_or_mode_fs_storage.res_config_settings_form_view
+msgid "Origin of fs storage for payment"
+msgstr "Origine du stockage pour le paiement"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model,name:account_payment_method_or_mode_fs_storage.model_account_payment_method
+msgid "Payment Methods"
+msgstr "Methodes de paiement"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model,name:account_payment_method_or_mode_fs_storage.model_account_payment_mode
+msgid "Payment Modes"
+msgstr "Modes de paiement"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model,name:account_payment_method_or_mode_fs_storage.model_account_payment_order
+msgid "Payment Order"
+msgstr "Ordre de paiement"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model_terms:ir.ui.view,arch_db:account_payment_method_or_mode_fs_storage.res_config_settings_form_view
+msgid "Select the storages allowed"
+msgstr "Sélectionner les stockages autorisés"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model_terms:ir.ui.view,arch_db:account_payment_method_or_mode_fs_storage.res_config_settings_form_view
+msgid "Select where the information about the storage is taken"
+msgstr "Sélectionner où l'information relative aux stockage est considérée"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model.fields,field_description:account_payment_method_or_mode_fs_storage.field_account_payment_method__storage
+#: model:ir.model.fields,field_description:account_payment_method_or_mode_fs_storage.field_account_payment_mode__storage
+msgid "Storage"
+msgstr "Stockage"
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model.fields,help:account_payment_method_or_mode_fs_storage.field_account_payment_method__storage
+#: model:ir.model.fields,help:account_payment_method_or_mode_fs_storage.field_account_payment_mode__storage
+msgid "Storage where to put the file after generation"
+msgstr "Stockage pour le fichier après génération"
+
+#. module: account_payment_method_or_mode_fs_storage
+#. odoo-python
+#: code:addons/account_payment_method_or_mode_fs_storage/models/account_payment_order.py:0
+#, python-format
+msgid "The file has been generated and dropped on the storage."
+msgstr "Le fichier a été généré et stocké"
+
+#. module: account_payment_method_or_mode_fs_storage
+#. odoo-python
+#: code:addons/account_payment_method_or_mode_fs_storage/models/account_payment_order.py:0
+#, python-format
+msgid ""
+"Unknown issue to upload the file on the storage:\n"
+"{details}"
+msgstr ""
+
+#. module: account_payment_method_or_mode_fs_storage
+#: model:ir.model.fields,field_description:account_payment_method_or_mode_fs_storage.field_res_company__fs_storage_source_payment
+#: model:ir.model.fields,field_description:account_payment_method_or_mode_fs_storage.field_res_config_settings__fs_storage_source_payment
+msgid "fs storage source payment"
+msgstr ""
