+++ account_payment_method_or_mode_fs_storage/security/res_groups.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2024 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo>
+  <record model="res.groups" id="group_fs_storage_manager">
+    <field name="name">Fs storage: Manager</field>
+  </record>
+</odoo>
