+++ account_payment_method_or_mode_fs_storage/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_account_payment_method_or_mode_fs_storage
