+++ account_payment_method_or_mode_fs_storage/views/account_payment_mode.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- Copyright 2024 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record model="ir.ui.view" id="account_payment_mode_form_view">
+        <field
+            name="name"
+        >account.payment.mode.form (in account_payment_method_or_mode_fs_storage)</field>
+        <field name="model">account.payment.mode</field>
+        <field name="inherit_id" ref="account_payment_mode.account_payment_mode_form" />
+        <field name="arch" type="xml">
+            <field name="payment_type" position="after">
+                <field name="storage" />
+            </field>
+        </field>
+    </record>
+
+
+
+</odoo>
