--- setup/account_payment_method_fs_storage/odoo/addons/account_payment_method_fs_storage
@@ -1,1 +1,0 @@
-../../../../account_payment_method_fs_storage
