+++ setup/account_payment_method_or_mode_fs_storage/odoo/addons/account_payment_method_or_mode_fs_storage
@@ -1,0 +1,1 @@
+../../../../account_payment_method_or_mode_fs_storage
